backend/requirements.txt
flask==2.3.3
python-dotenv==1.0.0
requests==2.31.0
pillow==10.0.1
openai==1.3.7
mysql-connector-python==9.3.0